q6voice: set per vocoder flag while set cal
In case when any client other than audio hal does acdb init
not all the flags set correctly, including per-vocoder flag.
When audio hal eventually does acdb init, it updates the flags
but that info is not propagated to voice driver.
Check for per-vocoder flag on voice cal set to check if new
flag is added and update accordingly.

Bug: 35950797
Signed-off-by: vivek mehta <mvivek@codeaurora.org>
Change-Id: I8c4cd91da7475645ba9fdbc5252193e649c62c2b